refactor(evo-animation) Move implementation of genealogical path visualizer to a separate file, enable its usage from OptHistory.show
