+++ golem/visualisation/opt_viz_extra.py
@@ -1,17 +1,17 @@
 import itertools
-import math
 import os
 from copy import deepcopy
 from datetime import datetime
-from functools import partial
 from glob import glob
 from os import remove
-from typing import Any, List, Sequence, Tuple, Optional, Callable
+from typing import Any, List, Sequence, Tuple, Optional
 
 import numpy as np
 import pandas as pd
 import seaborn as sns
-from matplotlib import pyplot as plt, animation
+from PIL import Image
+from imageio import get_writer, v2
+from matplotlib import pyplot as plt
 
 from golem.core.dag.graph import Graph
 from golem.core.log import default_log
@@ -19,8 +19,6 @@ from golem.core.optimisers.opt_history_objects.individual import Individual
 from golem.core.optimisers.opt_history_objects.opt_history import OptHistory
 from golem.core.paths import default_data_dir
 from golem.visualisation.graph_viz import GraphVisualizer
-from PIL import Image
-from imageio import get_writer, v2
 
 
 class OptHistoryExtraVisualizer:
@@ -193,81 +191,6 @@ class OptHistoryExtraVisualizer:
         plt.clf()
         plt.close('all')
 
-    def visualize_best_genealogical_path(self, graph_dist: Callable[[Graph, Graph], float] = None,
-                                         target_graph: Graph = None, save_as_gif=False):
-        """
-        Takes the best individual from the resultant generation and traces its genealogical path
-        taking the most similar parent each time (or the first parent if no similarity measure is provided).
-        That makes the picture more stable (and hence comprehensible) and the evolution process more apparent.
-
-        Saves the result as a GIF with the following layout:
-        - target graph (if provided) is displayed on the left,
-        - evolving graphs go as the next subplot, they evolve from the first generation to the last,
-        - and the fitness plot on the right shows fitness dynamics as the graphs evolve.
-        """
-        # Treating all graphs as equally distant if there's no reasonable way to compare them:
-        graph_dist = graph_dist or (lambda g1, g2: 1)
-
-        def draw_graph(graph: Graph, ax, title, highlight_title=False):
-            ax.clear()
-            ax.set_title(title, fontsize=22, color='green' if highlight_title else 'black')
-            self.graph_visualizer(graph).draw_nx_dag(ax, node_names_placement='legend')
-
-        last_internal_graph = self.history.archive_history[-1][0]
-        genealogical_path = trace_genealogical_path(last_internal_graph, graph_dist)
-
-        target_time_s = 5.
-        hold_result_time_s = 2.
-
-        figure_width = 5
-        width_ratios = [1.3, 0.7]
-        if target_graph is not None:
-            width_ratios = [1.3] + width_ratios
-
-        fig, axes = plt.subplots(
-            1, len(width_ratios),
-            figsize=(figure_width * sum(width_ratios), figure_width),
-            gridspec_kw={'width_ratios': width_ratios}
-        )
-        evo_ax, fitness_ax = axes[-2:]
-        if target_graph is not None:
-            draw_graph(target_graph, axes[0], "Target graph")  # Persists throughout the animation
-
-        fitnesses_along_path = list(map(lambda ind: ind.fitness.value, genealogical_path))
-        generations_along_path = list(map(lambda ind: ind.native_generation, genealogical_path))
-
-        def render_frame(frame_index):
-            path_index = min(frame_index, len(genealogical_path) - 1)
-            is_hold_stage = frame_index >= len(genealogical_path)
-
-            draw_graph(
-                genealogical_path[path_index].graph, evo_ax,
-                f"Evolution process,\ngeneration {generations_along_path[path_index]}/{generations_along_path[-1]}",
-                highlight_title=is_hold_stage
-            )
-            # Select only the genealogical path
-            fitness_ax.clear()
-            plot_fitness_with_axvline(
-                generations=generations_along_path,
-                fitnesses=fitnesses_along_path,
-                ax=fitness_ax,
-                axvline_x=generations_along_path[path_index],
-                current_fitness=fitnesses_along_path[path_index]
-            )
-            return evo_ax, fitness_ax
-
-        frames = len(genealogical_path) + \
-            int(math.ceil(len(genealogical_path) * hold_result_time_s / (hold_result_time_s + target_time_s)))
-        seconds_per_frame = (target_time_s + hold_result_time_s) / frames
-        fps = round(1 / seconds_per_frame)
-
-        anim = animation.FuncAnimation(fig, render_frame, repeat=False, frames=frames,
-                                       interval=1000 * seconds_per_frame)
-
-        if save_as_gif:
-            anim.save(os.path.join(self.save_path, "evolution_process.gif"), fps=fps)
-        plt.show()
-
 
 def visualise_pareto(front: Sequence[Individual],
                      objectives_numbers: Tuple[int, int] = (0, 1),
@@ -372,25 +295,3 @@ def objectives_lists(individuals: List[Any], objectives_numbers: Tuple[int] = No
     return objectives_values_set
 
 
-# Implementation details for genealogical path visualisation:
-def trace_genealogical_path(individual: Individual, graph_dist: Callable[[Graph, Graph], float]) -> List[Individual]:
-    # Choose nearest parent each time:
-    genealogical_path: List[Individual] = [individual]
-    while genealogical_path[-1].parents_from_prev_generation:
-        genealogical_path.append(max(
-            genealogical_path[-1].parents_from_prev_generation,
-            key=partial(graph_dist, genealogical_path[-1])
-        ))
-
-    return list(reversed(genealogical_path))
-
-
-def plot_fitness_with_axvline(generations: List[int], fitnesses: List[float], ax: plt.Axes, current_fitness: float,
-                              axvline_x: int = None):
-    ax.plot(generations, fitnesses)
-    ax.set_title(f'Metric dynamic,\ncurrent: {current_fitness}', fontsize=22)
-    ax.set_xlabel('Generation')
-    ax.set_ylabel('Metric')
-    if axvline_x is not None:
-        ax.axvline(x=axvline_x, color='black')
-    return ax
